refactor: Rename logic package to handler
