+++ cmd/server/handler/gameserver.go
@@ -1,4 +1,4 @@
-package logic
+package handler
 
 import (
 	"context"
